Update setup_package.py in utils package to use extension_helpers
